+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 
-## [ upcoming ]
+## [1.2.0] - 2026-02-18
 ### Features
 - **CodeLens**: Added "Run Script" CodeLens for `SET TERM` blocks and "Run Query" CodeLens for standard queries.